RecommenderSystem_docker/requirements.txt
azure-storage-blob 
azure-identity
pandas-tfrecords
numpy
pandas
matplotlib